requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
Pillow==10.3.0
pytesseract==0.3.13
PyMuPDF==1.24.11
transformers==4.45.2
requests==2.31.0
numpy==1.26.4
opencv-python==4.9.0.80
tqdm==4.66.2
tools==0.1.9
python-dotenv==1.0.1
pymongo==4.10.1
Levenshtein==0.26.1